consolidate test classes
