+++ service/src/test/java/bio/terra/externalcreds/services/ProviderServiceTest.java
@@ -16,11 +16,15 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import bio.terra.common.exception.BadRequestException;
+import bio.terra.common.exception.ForbiddenException;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.externalcreds.BaseTest;
 import bio.terra.externalcreds.ExternalCredsException;
 import bio.terra.externalcreds.TestUtils;
 import bio.terra.externalcreds.auditLogging.AuditLogEvent;
+import bio.terra.externalcreds.auditLogging.AuditLogEvent.Builder;
+import bio.terra.externalcreds.auditLogging.AuditLogEventType;
+import bio.terra.externalcreds.auditLogging.AuditLogger;
 import bio.terra.externalcreds.config.ExternalCredsConfig;
 import bio.terra.externalcreds.config.ProviderProperties;
 import bio.terra.externalcreds.dataAccess.GA4GHPassportDAO;
@@ -1109,5 +1113,128 @@ public class ProviderServiceTest extends BaseTest {
     return ClientRegistration.withRegistrationId(provider.toString())
         .authorizationGrantType(AuthorizationGrantType.REFRESH_TOKEN)
         .build();
+  }
+
+  @Nested
+  @TestComponent
+  class LinkedAccountTest {
+    @Autowired private ProviderService providerService;
+    @MockitoBean private AuditLogger auditLoggerMock;
+    @MockitoBean private LinkedAccountService linkedAccountService;
+    @MockitoBean private ProviderTokenClientCache providerTokenClientCacheMock;
+    @MockitoBean private OAuth2Service oAuth2ServiceMock;
+    @MockitoBean private JwtUtils jwtUtilsMock;
+
+    private final Provider provider = Provider.GITHUB;
+    private final String userId = UUID.randomUUID().toString();
+    private final String clientIP = "127.0.0.1";
+    private final AuditLogEvent.Builder auditLogEventBuilder =
+        new Builder().provider(provider).userId(userId).clientIP(clientIP);
+
+    @Test
+    void testLogLinkCreatePassportSuccess() {
+      when(jwtUtilsMock.getJwtTransactionClaim(anyString()))
+          .thenReturn(Optional.of("unit-test-claim"));
+
+      LinkedAccount linkedAccount = TestUtils.createRandomLinkedAccount(provider);
+      LinkedAccountWithPassportAndVisas linkedAccountWithPassportAndVisas =
+          new LinkedAccountWithPassportAndVisas.Builder()
+              .linkedAccount(linkedAccount)
+              .passport(TestUtils.createRandomPassport())
+              .build();
+
+      providerService.logLinkCreation(
+          Optional.of(linkedAccountWithPassportAndVisas), auditLogEventBuilder);
+      verify(auditLoggerMock)
+          .logEvent(
+              new AuditLogEvent.Builder()
+                  .auditLogEventType(AuditLogEventType.LinkCreated)
+                  .provider(provider)
+                  .userId(userId)
+                  .transactionClaim("unit-test-claim")
+                  .externalUserId(linkedAccount.getExternalUserId())
+                  .clientIP(clientIP)
+                  .build());
+    }
+
+    @Test
+    void testLogLinkNonPassportSuccess() {
+      LinkedAccount linkedAccount = TestUtils.createRandomLinkedAccount(provider);
+      LinkedAccountWithPassportAndVisas linkedAccountWithPassportAndVisas =
+          new LinkedAccountWithPassportAndVisas.Builder().linkedAccount(linkedAccount).build();
+
+      providerService.logLinkCreation(
+          Optional.of(linkedAccountWithPassportAndVisas), auditLogEventBuilder);
+      verify(auditLoggerMock)
+          .logEvent(
+              new AuditLogEvent.Builder()
+                  .auditLogEventType(AuditLogEventType.LinkCreated)
+                  .provider(provider)
+                  .userId(userId)
+                  .externalUserId(linkedAccount.getExternalUserId())
+                  .clientIP(clientIP)
+                  .build());
+    }
+
+    @Test
+    void testLogLinkCreateFailure() {
+      providerService.logLinkCreation(Optional.empty(), auditLogEventBuilder);
+      verify(auditLoggerMock)
+          .logEvent(
+              new AuditLogEvent.Builder()
+                  .auditLogEventType(AuditLogEventType.LinkCreationFailed)
+                  .provider(provider)
+                  .userId(userId)
+                  .externalUserId(Optional.empty())
+                  .clientIP(clientIP)
+                  .build());
+    }
+
+    @Test
+    void testGetProviderAccessTokenNoLinkedAccount() {
+      when(linkedAccountService.getLinkedAccount(userId, provider)).thenReturn(Optional.empty());
+
+      assertThrows(
+          NotFoundException.class,
+          () -> providerService.getProviderAccessToken(userId, provider, auditLogEventBuilder));
+    }
+
+    @Test
+    void testGetProviderAccessTokenExpiredLinkedAccount() {
+      var expiredLinkedAccount =
+          TestUtils.createRandomLinkedAccount(provider)
+              .withExpires(Timestamp.from(Instant.now().minusSeconds(60)));
+      when(linkedAccountService.getLinkedAccount(userId, provider))
+          .thenReturn(Optional.of(expiredLinkedAccount));
+
+      assertThrows(
+          ForbiddenException.class,
+          () -> providerService.getProviderAccessToken(userId, provider, auditLogEventBuilder));
+    }
+
+    @Test
+    void testGetProviderAccessTokenUnauthorized() {
+      var linkedAccount = TestUtils.createRandomLinkedAccount(provider);
+      var clientRegistration = TestUtils.createClientRegistration(linkedAccount.getProvider());
+
+      when(linkedAccountService.getLinkedAccount(linkedAccount.getUserId(), provider))
+          .thenReturn(Optional.of(linkedAccount));
+      when(providerTokenClientCacheMock.getProviderClient(linkedAccount.getProvider()))
+          .thenReturn(clientRegistration);
+      when(oAuth2ServiceMock.authorizeWithRefreshToken(
+              eq(clientRegistration),
+              eq(new OAuth2RefreshToken(linkedAccount.getRefreshToken(), null)),
+              any(Set.class)))
+          .thenThrow(
+              new OAuth2AuthorizationException(new OAuth2Error(OAuth2ErrorCodes.INVALID_TOKEN)));
+
+      var linkedAccountUserId = linkedAccount.getUserId();
+
+      assertThrows(
+          OAuth2AuthorizationException.class,
+          () ->
+              providerService.getProviderAccessToken(
+                  linkedAccountUserId, provider, auditLogEventBuilder));
+    }
   }
 }
